Refactor user detail output in 'look' command
This commit refactors the 'look' command in the game to bundle all the user detail display responses into one single string. This approach replaces the multiple 'userManager.send' calls, which were previously used for sending each of these responses separately. This refactor helps to improve the code readability and the efficiency of the messaging system.